refactor(macos): split manual AI activation flow
